pallet_allocations::pallet
Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

batch

Optimized allocation call, which will batch allocations of various amounts and destinations and together. This allow us to be much more efficient and thus increase our chain’s capacity in handling these transactions.

set_curve_starting_block

Warning: Doc-Only