Trait pallet_allocations::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Currency: Currency<Self::AccountId>;
    type PalletId: Get<PalletId>;
    type ProtocolFee: Get<Perbill>;
    type ProtocolFeeReceiver: WithAccountId<Self::AccountId>;
    type ExistentialDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type MaxAllocs: Get<u32>;
    type OracleMembers: Contains<Self::AccountId>;
    type MintCurve: Get<&'static MintCurve<Self>>;
    type BlockNumberProvider: BlockNumberProvider<BlockNumber = Self::BlockNumber>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Currency: Currency<Self::AccountId>

type PalletId: Get<PalletId>

type ProtocolFee: Get<Perbill>

type ProtocolFeeReceiver: WithAccountId<Self::AccountId>

type ExistentialDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

Runtime existential deposit

type MaxAllocs: Get<u32>

How big a batch can be

type OracleMembers: Contains<Self::AccountId>

type MintCurve: Get<&'static MintCurve<Self>>

MintCurve acts as an upper bound limiting how much the total token issuance can inflate over a configured session

type BlockNumberProvider: BlockNumberProvider<BlockNumber = Self::BlockNumber>

Provide access to the block number that should be used in mint curve calculations. For example those who use this pallet for a parachain may decide to use the block creation pace of the relay chain for timing.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.