Module pallet_allocations::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

BenchmarkOracles

Warning: Doc-Only

MintCurveStartingBlock

The block from which the mint curve should be considered starting its first inflation step

NextSessionQuota

The next session’s allocation quota, in other words, the top up that is coming for SessionQuota.

SessionQuota

The transitional allocation quota that is left for the current session.

SessionQuotaCalculationSchedule

The block in or after which the the session quota should be calculated

SessionQuotaRenewSchedule

The block in or after which the the session quota should be renewed

StorageVersion

Warning: Doc-Only